v1.14.1
-- Add stim toggle functionality to StimEditGUI
-- Fix matlab exception thrown by hmrR_BandpassFilt because of unexpected single type instead of double which the filtfilt function does not accept as argument. This was because I changed doubles to singles in NirsClass which when converted to snirf by homer3 generated singles in the datatimeseries in DataClass.

+++ FuncRegistry/UserFunctions/hmrR_BandpassFilt.m
@@ -53,7 +53,7 @@ for ii=1:length(data)
     else
         %    [fb,fa] = MakeFilter(FilterType,FilterOrder,fs,lpf,'low',Filter_Rp);
     end
-    ylpf=filtfilt(fb,fa,y);
+    ylpf = filtfilt(fb,fa,double(y));
     
     % high pass filter
     FilterType = 1;
